// lib/interfaces/workProject.ts
export interface IWorkProject {
    name: string
    summary: string
    preview: string
    description: string
    date: string
    tech: string[]
    projectType: string
    githubLink: string
    liveLink: string
    slug: string
    imageCover: string
    images: string[]
    previewOnHomepage: boolean
}